backend/requirements.txt
Flask
Flask-CORS
python-dotenv
pymongo
langchain
langchain-google-genai
langchain-community
faiss-cpu
sentence-transformers
numpy
requests